Запустил всё в Docker, небольшие изменения в для корректной работы Liquibase

+++ registration-service/registration-service-api/src/main/java/com/targetmonkey/registrationserviceapi/resource/v1/AdClientCompanyRegistrationV1.java
@@ -14,15 +14,15 @@ public interface AdClientCompanyRegistrationV1 {
     String saveAdCompany(@RequestBody AdCompanyRegistrationDTO adCompanyRegistrationDTO);
 
     @GetMapping("company/{id}")
-    AdCompanyRegistrationDTO getToId(@PathVariable long id);
+    AdCompanyRegistrationDTO getToId(@PathVariable int id);
 
     @GetMapping("all")
     List<AdCompanyRegistrationDTO> getAll();
 
     @PutMapping("edit/{id}")
-    String editAdCompany(@PathVariable long id,
+    String editAdCompany(@PathVariable int id,
                           @RequestBody AdCompanyRegistrationDTO adCompanyRegistrationDTO);
 
     @DeleteMapping("delete/{id}")
-    String deleteAdCompany(@PathVariable long id);
+    String deleteAdCompany(@PathVariable int id);
 }
